job_agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
openai>=1.0.0
google-generativeai>=0.7.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
sqlite3; sys_platform == "win32"
